bug in CAT, CSHIFT and parameter in toB1, toB2,.. now passed as value the parameter is no longer passed as refernce in order to allow expressions as output value
debug info in Character now with explicit cast to (int) for %d format
to avoid warnings

DationTests modified for behavior with empty field in GET and F-format

+++ openpearl-code/runtime/common/Character.h
@@ -128,7 +128,7 @@ namespace pearlrt {
          size_t i;
 
          if (length > 32767 || length < 1) {
-            printf("Character: illegal length (%d)\n", length);
+            printf("Character: illegal length (%d)\n", (int)length);
             throw theCharacterTooLongSignal;
          }
 
@@ -150,7 +150,7 @@ namespace pearlrt {
          unsigned int l = strlen(string);
 
          if (length > 32767 || l > length) {
-            printf("Character: illegal length (%d)\n", length);
+            printf("Character: illegal length (%d)\n", (int)length);
             throw theCharacterTooLongSignal;
          }
